test(extra): move vmconnect binding test
Keep the VMConnect library test harness disabled and run the native CredSSP binding vector from the shared extra-tier tests.
