+++ crates/ironrdp-vmconnect/src/native_credssp.rs
@@ -487,52 +487,10 @@ impl Drop for NativeSecurityContext {
     }
 }
 
-fn binding_hash(magic: &[u8], nonce: &[u8; NONCE_SIZE], public_key: &[u8]) -> Vec<u8> {
+pub(super) fn binding_hash(magic: &[u8], nonce: &[u8; NONCE_SIZE], public_key: &[u8]) -> Vec<u8> {
     let mut hasher = Sha256::new();
     hasher.update(magic);
     hasher.update(nonce);
     hasher.update(public_key);
     hasher.finalize().to_vec()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn binding_hash_matches_credssp_v5_v6_vector() {
-        // MS-CSSP 3.1.5: SHA256(magic || nonce || SubjectPublicKey), including the magic string's NUL.
-        let nonce = [
-            0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0A, 0x0B, 0x0C, 0x0D, 0x0E, 0x0F, 0x10, 0x11,
-            0x12, 0x13, 0x14, 0x15, 0x16, 0x17, 0x18, 0x19, 0x1A, 0x1B, 0x1C, 0x1D, 0x1E, 0x1F,
-        ];
-        let public_key = [1, 2, 3, 4];
-        let expected = [
-            0xB4, 0x4E, 0x3A, 0x42, 0x23, 0x29, 0x60, 0xBC, 0x17, 0x4B, 0x37, 0x7F, 0x77, 0xAE, 0xBF, 0xE4, 0xD6, 0x4B,
-            0xF7, 0x25, 0x41, 0x82, 0x62, 0xD8, 0x32, 0x84, 0x6B, 0xE4, 0x83, 0x29, 0x1F, 0x7B,
-        ];
-        assert_eq!(
-            binding_hash(CLIENT_SERVER_HASH_MAGIC, &nonce, &public_key).as_slice(),
-            expected
-        );
-
-        let mut changed_nonce = nonce;
-        changed_nonce[0] ^= 0x80;
-        assert_ne!(
-            binding_hash(CLIENT_SERVER_HASH_MAGIC, &changed_nonce, &public_key).as_slice(),
-            expected
-        );
-
-        let mut changed_public_key = public_key;
-        changed_public_key[0] ^= 0x80;
-        assert_ne!(
-            binding_hash(CLIENT_SERVER_HASH_MAGIC, &nonce, &changed_public_key).as_slice(),
-            expected
-        );
-
-        assert_ne!(
-            binding_hash(SERVER_CLIENT_HASH_MAGIC, &nonce, &public_key).as_slice(),
-            expected
-        );
-    }
 }
